created mobile and desktop chart for weight
