+++ apps/gym/src/app/(user)/weight/Weight.module.css
@@ -9,7 +9,7 @@
 
 .body {
   display: flex;
-  flex-flow: row;
+  flex-flow: row wrap;
   justify-content: space-between;
   flex-grow: 1;
   max-height: calc(100vh - 104px);
