+++ apps/gym/src/components/isolated/Weight/WeighChart/WeightChartDesktop/WeightChartDesktop.module.css
@@ -1,0 +1,24 @@
+.container {
+  display: flex;
+  flex-direction: column;
+  width: 70%;
+  max-height: 100%;
+}
+
+.chart_container {
+  display: flex;
+  border-radius: 20px;
+  width: calc(100% - 40px);
+  height: 100%;
+  flex-direction: column;
+  box-shadow: var(--box-shadow);
+  padding: 20px;
+  border-radius: 20px;
+  position: relative;
+}
+
+@media (width < 480px) {
+  .container {
+    display: none;
+  }
+}
