+++ apps/gym/src/components/isolated/Weight/WeighChart/WeightChartMobile/WeightChartMobile.module.css
@@ -1,0 +1,5 @@
+@media (width > 480px) {
+  .container {
+    display: none;
+  }
+}
