+++ apps/gym/src/components/isolated/Weight/WeighChart/WeightChartMobile/WeightChartMobile.tsx
@@ -1,0 +1,6 @@
+import styles from './WeightChartMobile.module.css'
+import { WeightChartMobileProps } from '@/types'
+
+export const WeightChartMobile = ({}: WeightChartMobileProps) => {
+  return <div className={styles.container}>WeightChartMobile</div>
+}
